--- app/src/main/java/com/techcourse/dao/UpdateJdbcTemplate.java
@@ -1,45 +1,0 @@
-package com.techcourse.dao;
-
-import com.techcourse.domain.User;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-import javax.sql.DataSource;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-public class UpdateJdbcTemplate {
-
-    private static final Logger log = LoggerFactory.getLogger(UpdateJdbcTemplate.class);
-
-    private final DataSource dataSource;
-
-    public UpdateJdbcTemplate(final DataSource dataSource) {
-        this.dataSource = dataSource;
-    }
-
-    public void update(final User user) {
-        final var sql = createQueryForUpdate();
-
-        try (Connection conn = dataSource.getConnection();
-             PreparedStatement pstmt = conn.prepareStatement(sql)) {
-
-            log.debug("query : {}", sql);
-
-            setValuesForUpdate(user, pstmt);
-            pstmt.executeUpdate();
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            throw new RuntimeException(e);
-        }
-    }
-
-    private String createQueryForUpdate() {
-        return "update users set password = ? where id = ?";
-    }
-
-    private void setValuesForUpdate(User user, PreparedStatement pstmt) throws SQLException {
-        pstmt.setObject(1, user.getPassword());
-        pstmt.setObject(2, user.getId());
-    }
-}
